Almost
Queda el Async Function, Try Catch Finally, cambiar los colores, etc

+++ js/carrito.js
@@ -1,5 +1,5 @@
 // llamamo al localStorage
-const productosEnCarrito = JSON.parse(localStorage.getItem("productos-en-carrito"));
+let productosEnCarrito = localStorage.getItem("productos-en-carrito");
 productosEnCarrito = JSON.parse(productosEnCarrito);
 
 // creamos todas las constantes con sus querySelectors 
@@ -7,35 +7,37 @@ productosEnCarrito = JSON.parse(productosEnCarrito);
 const contenedorCarritoVacio = document.querySelector("#carrito-vacio");
 const contenedorCarritoProductos = document.querySelector("#carrito-productos");
 const contenedorCarritoAcciones = document.querySelector("#carrito-acciones");
-const contenedorCarritoComprando = document.querySelector("#carrito-comprado");
-let botonesEliminar = document.querySelectorAll(".carrito-producto-eliminar"); 
+const contenedorCarritoComprado = document.querySelector("#carrito-comprado");
+let botonesEliminar = document.querySelectorAll(".carrito-producto-eliminar");
 // este let lo usamos para eliminar y agarramos el querySelectorAll para justamente poder eliminar ALL
 const botonVaciar = document.querySelector("#carrito-acciones-vaciar");
 const contenedorTotal = document.querySelector("#total");
 const botonComprar = document.querySelector("#carrito-acciones-comprar");
 
 
-
 function cargarProductosCarrito() {
-    // si hay productos en carrito los mostramos y sino mostramos el mensaje
+        // si hay productos en carrito los mostramos y sino mostramos el mensaje
 
-if (productosEnCarrito && productosEnCarrito.length > 0) {
+    if (productosEnCarrito && productosEnCarrito.length > 0) {
+
+        contenedorCarritoVacio.classList.add("disabled");
+        contenedorCarritoProductos.classList.remove("disabled");
+        contenedorCarritoAcciones.classList.remove("disabled");
+        contenedorCarritoComprado.classList.add("disabled");
+
+        // esto lo hacemos para que quede el carrito productos totalmente vacio
     
-    contenedorCarritoVacio.classList.add("disabled");
-    contenedorCarritoProductos.classList.remove("disabled");
-    contenedorCarritoAcciones.classList.remove("disabled");
-    contenedorCarritoComprando.classList.add("disabled");
+        contenedorCarritoProductos.innerHTML = "";
 
+        // Creamos un forEach para que por cada producto creamos un div con la class CarritoProducto
+        // y el InnerHtml para poner todo lo que queremos en cada producto
 
-    // esto lo hacemos para que quede el carrito productos totalmente vacio
-    contenedorCarritoProductos.innerHTML = "";
-
-// Creamos un forEach para que por cada producto creamos un div con la class CarritoProducto
-// y el InnerHtml para poner todo lo que queremos en cada producto
-
-    productosEnCarrito.forEach(producto => {
-        div.classList.add("carrito-producto");
-        div.innerHTML = `
+    
+        productosEnCarrito.forEach(producto => {
+    
+            const div = document.createElement("div");
+            div.classList.add("carrito-producto");
+            div.innerHTML = `
                 <img class="carrito-producto-imagen" src="${producto.imagen}" alt="${producto.titulo}">
                 <div class="carrito-producto-titulo">
                     <small>Título</small>
@@ -56,46 +58,63 @@ if (productosEnCarrito && productosEnCarrito.length > 0) {
                 <button class="carrito-producto-eliminar" id="${producto.id}"><i class="bi bi-trash-fill"></i></button>
             `;
             // todo esto queremos que se haga un append en contenedorCarritoProductos
-
             contenedorCarritoProductos.append(div);
-    })
-
+        })
+    
     actualizarBotonesEliminar();
     actualizarTotal();
+	
+    } else {
+            // esto lo hacemos ya que queremos tambien que si borramos productos que tambien esto pase 
 
-}else{
-    // esto lo hacemos ya que queremos tambien que si borramos productos que tambien esto pase 
-
-    contenedorCarritoVacio.classList.remove("disabled");
-    contenedorCarritoProductos.classList.add("disabled");
-    contenedorCarritoAcciones.classList.add("disabled");
-    contenedorCarritoComprando.classList.add("disabled");
-
+        contenedorCarritoVacio.classList.remove("disabled");
+        contenedorCarritoProductos.classList.add("disabled");
+        contenedorCarritoAcciones.classList.add("disabled");
+        contenedorCarritoComprado.classList.add("disabled");
     }
-    
+
 }
-
-
 
 cargarProductosCarrito();
 
 
-// Creamos las distintas funcioens para eliminar del carrito y actualizar del carrito 
+// Creamos las distintas funciones para eliminar del carrito y actualizar del carrito 
 
 function actualizarBotonesEliminar() {
-     botonesEliminar = document.querySelectorAll(".producto-eliminar");
+    botonesEliminar = document.querySelectorAll(".carrito-producto-eliminar");
 
-        botonesEliminar.forEach(boton => {
+    botonesEliminar.forEach(boton => {
         boton.addEventListener("click", eliminarDelCarrito);
     });
 }
 
 
 
-function eliminarDelCarrito(){
+
+function eliminarDelCarrito(e) {
+    Toastify({
+        text: "Producto eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background: "linear-gradient(to right, #4b33a8, #785ce9)",
+          borderRadius: "2rem",
+          textTransform: "uppercase",
+          fontSize: ".75rem"
+        },
+        offset: {
+            x: '1.5rem', // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: '1.5rem' // vertical axis - can be a number or a string indicating unity. eg: '2em'
+          },
+        onClick: function(){} // Callback after click
+      }).showToast();
+
     const idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
-
+    
     productosEnCarrito.splice(index, 1);
     cargarProductosCarrito();
 
@@ -103,16 +122,27 @@ function eliminarDelCarrito(){
 
 }
 
+// le agregamos un addEventListener para llamar a la funcion de vaciar  e implementamos su libreria
 
-
-// le agregamos un addEventListener para llamar a la funcion de vaciar 
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 function vaciarCarrito() {
 
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+    Swal.fire({
+        title: '¿Estás seguro?',
+        icon: 'question',
+        html: `Se van a borrar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+      })
 }
 
 
@@ -124,8 +154,24 @@ para justamente la suma de ellos
 Y hacemos el boton nuevamente como hicimos arriba
 */
 
+
+
 function actualizarTotal() {
-    const totalCalculado = productosEnCarrito.reduce((acc, producto) => + (producto.precio * producto.cantidad), 0);
+    const totalCalculado = productosEnCarrito.reduce((acc, producto) => acc + (producto.precio * producto.cantidad), 0);
     total.innerText = `$${totalCalculado}`;
+}
+
+// volvemos a utilizar el addEventListener
+
+botonComprar.addEventListener("click", comprarCarrito);
+function comprarCarrito() {
+
+    productosEnCarrito.length = 0;
+    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+    
+    contenedorCarritoVacio.classList.add("disabled");
+    contenedorCarritoProductos.classList.add("disabled");
+    contenedorCarritoAcciones.classList.add("disabled");
+    contenedorCarritoComprado.classList.remove("disabled");
 
 }
